Refactored the logout module
