+++ src/app/logout/logout-routing.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import {LogoutComponent} from './logout/logout.component';
+import {JwtGuardGuard} from '../guards/jwt-guard.guard';
+
+
+const routes: Routes = [
+  {path: 'logout', component: LogoutComponent, canActivate: [JwtGuardGuard]}
+];
+
+@NgModule({
+  imports: [RouterModule.forRoot(routes)],
+  exports: [RouterModule]
+})
+export class LogoutRoutingModule { }
