+++ src/app/logout/logout/logout.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {AuthServiceService} from '../services/auth-service.service';
+import {AuthServiceService} from '../../services/auth-service.service';
 import { Router } from '@angular/router';
 
 @Component({
